Template 삭제 시 TemplateDoc에서 해당 TemplateId를 가진 레코드 전부 삭제하는 기능 컨트롤러 / 서비스 / 레포지토리 구현

+++ src/main/java/com/kns/tenquest/service/TemplateService.java
@@ -62,7 +62,13 @@ public class TemplateService {
         return templateRepository.findById(templateId).get();
     }
 
-    public  void templateDelete(String templateId){
-        templateRepository.deleteById(templateId);
+    public  int templateDelete(String templateId){
+        Optional<Template> optTemplateDto = templateRepository.findById(templateId);
+        if(optTemplateDto.isPresent()){
+            templateRepository.deleteById(templateId);
+            return ResponseStatus.OK.getCode();
+        }
+        return ResponseStatus.NOT_FOUND.getCode();
+
     }
 }
